NLPbackend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-socketio==5.10.0
python-engineio==4.8.0
websockets==11.0.3
deep-translator==1.11.4
pydantic==2.4.2
aiohttp==3.8.6